Extract Codex CLI probe markdown renderers
